requirements.txt
pandas
requests
tqdm
scipy
python-dotenv
